Vid: Update Item
problem: when dragging a listEl
~~~~~~
DOMException: Failed to execute 'appendChild' on 'Node': The node to be removed is no longer a child of this node. Perhaps it was moved in a 'blur' event handler?

this happened because when you click on a certain listEl, it's in edit mode, so if you move it in that state, it will give you that error

solution:
we want to only be able to update our items in a certain condition
that is.. if we are not dragging it

-- added a global variable called let dragging = false
 (line 24)
-- modified: dragging = true while dragging (line 179)
-- modified: dragging = false once dropped (line 204)
-- then added a conditional statement in the updateItem() (line 120-127)